tcc: provide implementations for some non-called functions
TCC is not an optimizing compiler, so it can't easily see through
some constructs used in the linux kernel that aren't standard conforming.

For instance such idiom (for ext4_decrypt):

   bool used = false;
   if (some_function_call (args))
     used = true;
   if (used)
     ext4_decrypt(...);

If ext4 encryption is configured off then no implementation
of ext4_decrypt is provided at all, and some_function_call
will be often an trivial inline function returning zero.
But even if (say) defined to zero TCC doesn't know that 'used'
is only ever set to false, so it emits the call to ext4_decrypt,
and hence linking fails.

At runtime, of course, that call will never be actually done, because
'used' will be false.  So instead of hacking several TCC specific
places to #ifdef out also calls to such functions, instead
provide trivial implementations for those that simply BUG().

Also provide __builtin_ctzll.

This is one of the first hacks, so some of these symbols might
not be needed anymore.