tcc: hack around various limitations
TCC can't use a forward declaration of a type to declare a variable.
It's standard C, but doesn't occur in practice often (only case
seen so far), so let's just swap type and var decl in iommu.c .

TCCs assembler doesn't understand the "rex64/opcode" pseudo op.
Hack uses of it out for now and replace with a WARN_ON_CPU
(used currently only for fxsave/fxrstr when fxsaveq isn't available,
which it is also for TCC.

TCCs assembler doesn't support the use of 'inl (%dx)' with
parentheses.  It just wants 'inl %dx', which also works with GAS
(and arguably is more correct, as that operand is no memory address)