tcc: percpu changes
two bugfixes and one hack: in percpu_cmpxchg16b_double TCC would
need 8 registers, but it only has 7 (isn't as optimal as GCC).
Some of the asm operands aren't even used in the asm itself,
so we can simply remove them from the asm (and make sure we evaluate it
for side-effects).  That's the hack.

In x86_this_cpu_variable_test_bit it's argument 1 that is the
per-cpu memory, not argument 2.  It worked with GAS because of laxer
parsing.

In percpu_stable_op the uses of "%P1" template characters require
a "m"(var) argument, not a "p"(&var) argument.  It happens to work
with GCC due to better optimizers so it always sees the global
variable associated with 'var'.  If that were more convoluted
and the address would end up in a register (as with TCC) GAS would
complain (as does TCCs assembler) for invalid addresses like
"%gs:%rax" (it would have to be "%gs:(%rax)" ).  Simply use the
same way as in percpu_from_op.