Don't use compression in bzImage
with some configs and scripts to compile the kernel
specially the time for compressing and relinking the
kernel is actually longer than for compiling itself.
This hacks out all compression, it directly embeds the
ELF image in bzImage.  It still has to move
around the parts at boot time because of the ELF header and
page tables, so it's a bit slower than necessary during
boot (but still faster than with compression).