tcc: Add multiboot header and glue
this adds a multiboot header and associated glue
for transforming MB info into native linux info
(for the memory map), as well as 32bit startup code
to switch to 64 bit.

This means the vmlinux binary can be directly loaded by qemu,
obviating the need for 16bit code, and all the embedding/compressing
stuff.

Depending on the qemu version this requires some patches to it,
some newer ones like to complain about the MB header not being part
of the loadable image.  The MB code/data is hardcoded in a way
that it expects 2MB page sizes in the ELF image, so we hardcode
that for now as well.